refactor (extract interfaces) : Junit없이 테스트 가능하도록 인터페이스 추출
