+++ src/main/java/kitchenpos/bo/MenuGroupBo.java
@@ -1,6 +1,6 @@
 package kitchenpos.bo;
 
-import kitchenpos.dao.MenuGroupDao;
+import kitchenpos.dao.Interface.MenuGroupDao;
 import kitchenpos.model.MenuGroup;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
