+++ src/main/java/kitchenpos/dao/DefaultOrderDao.java
@@ -1,6 +1,7 @@
 package kitchenpos.dao;
 
-import kitchenpos.model.OrderTable;
+import kitchenpos.dao.Interface.OrderDao;
+import kitchenpos.model.Order;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
@@ -12,19 +13,20 @@ import org.springframework.stereotype.Repository;
 import javax.sql.DataSource;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 
 @Repository
-public class OrderTableDao {
-    private static final String TABLE_NAME = "order_table";
+public class DefaultOrderDao implements OrderDao {
+    private static final String TABLE_NAME = "orders";
     private static final String KEY_COLUMN_NAME = "id";
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert jdbcInsert;
 
-    public OrderTableDao(final DataSource dataSource) {
+    public DefaultOrderDao(final DataSource dataSource) {
         jdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
         jdbcInsert = new SimpleJdbcInsert(dataSource)
                 .withTableName(TABLE_NAME)
@@ -32,7 +34,8 @@ public class OrderTableDao {
         ;
     }
 
-    public OrderTable save(final OrderTable entity) {
+    @Override
+    public Order save(final Order entity) {
         if (Objects.isNull(entity.getId())) {
             final SqlParameterSource parameters = new BeanPropertySqlParameterSource(entity);
             final Number key = jdbcInsert.executeAndReturnKey(parameters);
@@ -42,7 +45,8 @@ public class OrderTableDao {
         return entity;
     }
 
-    public Optional<OrderTable> findById(final Long id) {
+    @Override
+    public Optional<Order> findById(final Long id) {
         try {
             return Optional.of(select(id));
         } catch (final EmptyResultDataAccessException e) {
@@ -50,50 +54,53 @@ public class OrderTableDao {
         }
     }
 
-    public List<OrderTable> findAll() {
-        final String sql = "SELECT id, table_group_id, number_of_guests, empty FROM order_table";
+    @Override
+    public List<Order> findAll() {
+        final String sql = "SELECT id, order_table_id, order_status, ordered_time FROM orders";
         return jdbcTemplate.query(sql, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    public List<OrderTable> findAllByIdIn(final List<Long> ids) {
-        final String sql = "SELECT id, table_group_id, number_of_guests, empty FROM order_table WHERE id IN (:ids)";
+    @Override
+    public boolean existsByOrderTableIdAndOrderStatusIn(final Long orderTableId, final List<String> orderStatuses) {
+        final String sql = "SELECT CASE WHEN COUNT(*) > 0 THEN TRUE ELSE FALSE END" +
+                " FROM orders WHERE order_table_id = (:orderTableId) AND order_status IN (:orderStatuses)";
         final SqlParameterSource parameters = new MapSqlParameterSource()
-                .addValue("ids", ids);
-        return jdbcTemplate.query(sql, parameters, (resultSet, rowNumber) -> toEntity(resultSet));
+                .addValue("orderTableId", orderTableId)
+                .addValue("orderStatuses", orderStatuses);
+        return jdbcTemplate.queryForObject(sql, parameters, Boolean.class);
     }
 
-    public List<OrderTable> findAllByTableGroupId(final Long tableGroupId) {
-        final String sql = "SELECT id, table_group_id, number_of_guests, empty" +
-                " FROM order_table WHERE table_group_id = (:tableGroupId)";
+    @Override
+    public boolean existsByOrderTableIdInAndOrderStatusIn(final List<Long> orderTableIds, final List<String> orderStatuses) {
+        final String sql = "SELECT CASE WHEN COUNT(*) > 0 THEN TRUE ELSE FALSE END" +
+                " FROM orders WHERE order_table_id IN (:orderTableIds) AND order_status IN (:orderStatuses)";
         final SqlParameterSource parameters = new MapSqlParameterSource()
-                .addValue("tableGroupId", tableGroupId);
-        return jdbcTemplate.query(sql, parameters, (resultSet, rowNumber) -> toEntity(resultSet));
+                .addValue("orderTableIds", orderTableIds)
+                .addValue("orderStatuses", orderStatuses);
+        return jdbcTemplate.queryForObject(sql, parameters, Boolean.class);
     }
 
-    private OrderTable select(final Long id) {
-        final String sql = "SELECT id, table_group_id, number_of_guests, empty FROM order_table WHERE id = (:id)";
+    private Order select(final Long id) {
+        final String sql = "SELECT id, order_table_id, order_status, ordered_time FROM orders WHERE id = (:id)";
         final SqlParameterSource parameters = new MapSqlParameterSource()
                 .addValue("id", id);
         return jdbcTemplate.queryForObject(sql, parameters, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    private void update(final OrderTable entity) {
-        final String sql = "UPDATE order_table SET table_group_id = (:tableGroupId)," +
-                " number_of_guests = (:numberOfGuests), empty = (:empty) WHERE id = (:id)";
+    private void update(final Order entity) {
+        final String sql = "UPDATE orders SET order_status = (:orderStatus) WHERE id = (:id)";
         final SqlParameterSource parameters = new MapSqlParameterSource()
-                .addValue("tableGroupId", entity.getTableGroupId())
-                .addValue("numberOfGuests", entity.getNumberOfGuests())
-                .addValue("empty", entity.isEmpty())
+                .addValue("orderStatus", entity.getOrderStatus())
                 .addValue("id", entity.getId());
         jdbcTemplate.update(sql, parameters);
     }
 
-    private OrderTable toEntity(final ResultSet resultSet) throws SQLException {
-        final OrderTable entity = new OrderTable();
+    private Order toEntity(final ResultSet resultSet) throws SQLException {
+        final Order entity = new Order();
         entity.setId(resultSet.getLong(KEY_COLUMN_NAME));
-        entity.setTableGroupId(resultSet.getObject("table_group_id", Long.class));
-        entity.setNumberOfGuests(resultSet.getInt("number_of_guests"));
-        entity.setEmpty(resultSet.getBoolean("empty"));
+        entity.setOrderTableId(resultSet.getLong("order_table_id"));
+        entity.setOrderStatus(resultSet.getString("order_status"));
+        entity.setOrderedTime(resultSet.getObject("ordered_time", LocalDateTime.class));
         return entity;
     }
 }
