+++ src/main/java/kitchenpos/dao/Interface/MenuDao.java
@@ -1,0 +1,16 @@
+package kitchenpos.dao.Interface;
+
+import kitchenpos.model.Menu;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface MenuDao {
+    Menu save(Menu entity);
+
+    Optional<Menu> findById(Long id);
+
+    List<Menu> findAll();
+
+    long countByIdIn(List<Long> ids);
+}
