+++ src/main/java/kitchenpos/dao/Interface/MenuGroupDao.java
@@ -1,0 +1,16 @@
+package kitchenpos.dao.Interface;
+
+import kitchenpos.model.MenuGroup;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface MenuGroupDao {
+    MenuGroup save(MenuGroup entity);
+
+    Optional<MenuGroup> findById(Long id);
+
+    List<MenuGroup> findAll();
+
+    boolean existsById(Long id);
+}
