+++ src/main/java/kitchenpos/dao/Interface/MenuProductDao.java
@@ -1,0 +1,14 @@
+package kitchenpos.dao.Interface;
+
+import kitchenpos.model.MenuProduct;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface MenuProductDao {
+    MenuProduct save(MenuProduct entity);
+
+    Optional<MenuProduct> findById(Long id);
+
+    List<MenuProduct> findAllByMenuId(Long menuId);
+}
