+++ src/main/java/kitchenpos/dao/Interface/OrderDao.java
@@ -1,0 +1,18 @@
+package kitchenpos.dao.Interface;
+
+import kitchenpos.model.Order;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface OrderDao {
+    Order save(Order entity);
+
+    Optional<Order> findById(Long id);
+
+    List<Order> findAll();
+
+    boolean existsByOrderTableIdInAndOrderStatusIn(List<Long> orderTableIds, List<String> orderStatuses);
+
+    boolean existsByOrderTableIdAndOrderStatusIn(Long orderTableId, List<String> orderStatuses);
+}
