+++ src/main/java/kitchenpos/dao/Interface/OrderLineItemDao.java
@@ -1,0 +1,16 @@
+package kitchenpos.dao.Interface;
+
+import kitchenpos.model.OrderLineItem;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface OrderLineItemDao {
+    OrderLineItem save(OrderLineItem entity);
+
+    Optional<OrderLineItem> findById(Long id);
+
+    List<OrderLineItem> findAll();
+
+    List<OrderLineItem> findAllByOrderId(Long orderId);
+}
