+++ src/main/java/kitchenpos/dao/Interface/OrderTableDao.java
@@ -1,0 +1,18 @@
+package kitchenpos.dao.Interface;
+
+import kitchenpos.model.OrderTable;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface OrderTableDao {
+    OrderTable save(OrderTable entity);
+
+    Optional<OrderTable> findById(Long id);
+
+    List<OrderTable> findAll();
+
+    List<OrderTable> findAllByIdIn(List<Long> ids);
+
+    List<OrderTable> findAllByTableGroupId(Long tableGroupId);
+}
