+++ src/main/java/kitchenpos/dao/Interface/TableGroupDao.java
@@ -1,0 +1,7 @@
+package kitchenpos.dao.Interface;
+
+import kitchenpos.model.TableGroup;
+
+public interface TableGroupDao {
+    TableGroup save(TableGroup entity);
+}
